+++ circuitx/navigation/src/jvmTest/kotlin/com/slack/circuitx/navigation/intercepting/TestUtils.kt
@@ -7,16 +7,14 @@ import com.slack.circuit.internal.test.TestScreen
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.screen.Screen
 import com.slack.circuit.test.FakeNavigator
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 
 internal val FailConsumed = InterceptedResult.Failure(consumed = true)
 internal val FailUnconsumed = InterceptedResult.Failure(consumed = false)
 internal val SuccessUnconsumed = InterceptedResult.Success(consumed = false)
 
 internal fun ComposeContentTestRule.setTestContent(
-  interceptors: ImmutableList<NavigationInterceptor> = persistentListOf(),
-  eventListeners: ImmutableList<NavigationEventListener> = persistentListOf(),
+  interceptors: List<NavigationInterceptor> = emptyList(),
+  eventListeners: List<NavigationEventListener> = emptyList(),
   notifier: InterceptingNavigator.FailureNotifier? = null,
   initialScreen: Screen = TestScreen.RootAlpha,
   additionalScreens: Array<Screen> = arrayOf(),